servers/pyproject.toml
[project]
name = "servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=0.6.3",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.4",
    "mcp[cli]>=1.7.1",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]